downloader/requirements.txt
kafka-python==2.0.2
SQLAlchemy==1.4.25
loguru==0.5.3
fire==0.4.0
pydantic==1.8.1
pytube==11.0.1
opencv-python==4.5.3.56
pytesseract==0.3.8
pandas==1.3.3
psycopg2-binary==2.9.1
